backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["fanfoo <wtr.ssmt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
openai = "^1.40.2"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
